added candlestick entity, fix influxdb config

+++ src/main/java/com/jtradeplatform/saas/currencyPair/CurrencyPairRepository.java
@@ -2,5 +2,5 @@ package com.jtradeplatform.saas.currencyPair;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CurrencyPairRepository extends JpaRepository<CurrencyPair, Long> {
+public interface CurrencyPairRepository extends JpaRepository<CurrencyPair, Integer> {
 }
